requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
audioread==3.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
datasets==3.3.2
decorator==5.2.1
dill==0.3.8
filelock==3.17.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pooch==1.8.2
propcache==0.3.0
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
